docs: dir structure
