+++ Army/Attack.hpp
