+++ Army/Soldier.hpp
